Fix font size type to match `impl Into<Pixels>`

+++ ui/src/pages/dashboard.rs
@@ -14,7 +14,7 @@ use crate::{
     themes::AppTheme,
 };
 
-const TITLE_FONT_SIZE: u16 = 22;
+const TITLE_FONT_SIZE: u32 = 22;
 const SAMPLE_EVERY: Duration = Duration::from_millis(1000);
 const TITLE_FONT: Font = Font {
     family: font::Family::Name("Noto Sans"),
